streamlined workouts, fetching and viewing them
