--- Olympsis/Components/Activities/WorkoutMapView.swift
@@ -1,22 +1,0 @@
-//
-//  RunningMapView.swift
-//  Olympsis
-//
-//  Created by Joel Joseph on 7/6/24.
-//
-
-import MapKit
-import SwiftUI
-
-struct WorkoutMapView: View {
-    var body: some View {
-        Map()
-            .clipShape(RoundedRectangle(cornerRadius: 10))
-            .padding(.horizontal)
-            .frame(height: 160)
-    }
-}
-
-#Preview {
-    WorkoutMapView()
-}
