--- Shared/Data/HKWorkoutStatistics.swift
@@ -1,8 +1,0 @@
-//
-//  HKWorkoutStatistics.swift
-//  Olympsis
-//
-//  Created by Joel Joseph on 7/23/24.
-//
-
-import Foundation
