+++ Shared/Data/WorkoutManager+Write.swift
@@ -1,0 +1,10 @@
+//
+//  WorkoutManager+Write.swift
+//  Olympsis
+//
+//  Created by Joel Joseph on 8/3/24.
+//
+
+import os
+import SwiftUI
+
